api/requirements.txt
flask==3.1.2
google-genai==1.57.0
flask-cors==6.0.2